apply feedback from Sam
